requirements.txt
openai==1.57.4
playsound==1.2.2
python-dotenv==1.0.1
